Code refactoring - extract MenuDrawer
